+++ R/addChar.R
@@ -106,14 +106,16 @@ addLocationChar <- function(clinChar, locationTable) {
 #' @param visitDetailTable a visitDetailTable object set using the the makeLocTable fn
 #' @param timeWindows a timeWindow object that specifies the boundaries relative to the target start date
 #' on when to search for the presence of a value. use `makeTimeTable` function
+#' @param asCount a toggle to determine whether to summarize as a count or presence
 #' @return adds a visitDetailChar object into the clinChar extractSettings slot
 #' @export
-addSpecialtyChar <- function(clinChar, visitDetailTable, timeWindows) {
+addSpecialtyChar <- function(clinChar, visitDetailTable, timeWindows, asCount) {
 
   specChar <- new("visitDetailChar", orderId = set_order_id(clinChar),
                   domain = "provider",
               visitDetailTable = visitDetailTable,
-              categoryId = 9001L)
+              categoryId = 9001L,
+              count = asCount)
 
   specChar@time <- timeWindows
   specChar@tempTables <- list(